+++ test/time_stack_item_test.rb
@@ -296,4 +296,22 @@ class TestTimeStackItem < Minitest::Test
       assert_equal dt, now, "#{dt.to_f}, #{now.to_f}"
     end
   end
+
+  def test_travel_offset_aligns_to_clock
+    t = Time.now
+    stack_item = Timecop::TimeStackItem.new(:travel, t)
+    travel_offset_denom = stack_item.travel_offset.to_r.denominator
+    clock_resolution = Process.clock_getres(:CLOCK_REALTIME, :hertz)
+    assert_equal 0, clock_resolution.modulo(travel_offset_denom),
+      "travel offset precision (#{travel_offset_denom}) does not align with clock resolution (#{clock_resolution})"
+  end
+
+  def test_travel_offset_aligns_to_travel_time
+    t = Time.now + 0.001_002_003_004
+    stack_item = Timecop::TimeStackItem.new(:travel, t)
+    travel_offset_denom = stack_item.travel_offset.to_r.denominator
+    travel_time_denom = t.to_r.denominator
+    assert_equal 0, travel_time_denom.modulo(travel_offset_denom),
+      "travel offset precision (#{travel_offset_denom}) does not align with travel time precision (#{travel_time_denom})"
+  end
 end